requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
selenium==4.15.2
webdriver-manager==4.0.1
aiohttp==3.9.1
